Move folding.py out of spyder.api to sourcecode.api and sourcecode.folding
