+++ spyder/widgets/sourcecode/codeeditor.py
@@ -62,7 +62,7 @@ from spyder.widgets.panels.edgeline import EdgeLine
 from spyder.widgets.panels.scrollflag import ScrollFlagArea
 from spyder.widgets.panels.manager import PanelsManager
 from spyder.widgets.panels.codefolding import FoldingPanel
-from spyder.api.folding import IndentFoldDetector
+from spyder.widgets.sourcecode.folding import IndentFoldDetector
 from spyder.widgets.sourcecode.utils.decoration import TextDecorationsManager
 from spyder.api.panel import Panel
